feat: introduce Vite and split frontend into ES modules (#28)
Replaces the monolithic 830-line app.js with a proper ES module
structure under frontend/src/. Adds Vite as the build tool with a
dev server proxy, updates Dockerfile.web to a two-stage Node+nginx
build, and retires public/ as a source directory.
