+++ Dockerfile.web
@@ -1,3 +1,12 @@
+# ── Build stage ─────────────────────────────────────────────
+FROM node:22-alpine AS build
+WORKDIR /frontend
+COPY frontend/package.json frontend/package-lock.json ./
+RUN npm ci
+COPY frontend/ .
+RUN npm run build
+
+# ── Serve stage ──────────────────────────────────────────────
 FROM nginx:alpine
-COPY public/ /usr/share/nginx/html
+COPY --from=build /frontend/dist /usr/share/nginx/html
 COPY nginx.conf /etc/nginx/conf.d/default.conf
