+++ bgci.code-workspace
@@ -1,0 +1,10 @@
+{
+	"folders": [
+		{
+			"path": "node/bgg-writer"
+		},
+		{
+			"path": "frontend"
+		}
+	]
+}
