+++ frontend/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "bgci-frontend",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^6.3.4"
+  }
+}
