+++ frontend/public/config.js.example
