+++ frontend/src/api.js
@@ -1,0 +1,57 @@
+import { setAllGames } from './state.js';
+
+const API = window.API_OVERRIDE ?? '/api';
+export { API };
+
+export async function loadConfig() {
+	try {
+		const res = await fetch(`${API}/config`);
+		return await res.json();
+	} catch {
+		return { bggConfigured: false, bggReachable: false, bggSearchEnabled: false, bggSyncEnabled: false };
+	}
+}
+
+export async function loadGames(onLoaded) {
+	const res = await fetch(`${API}/games`);
+	const games = await res.json();
+	setAllGames(games);
+	if (onLoaded) onLoaded();
+}
+
+export async function saveGame(payload, id) {
+	const url    = id ? `${API}/games/${id}` : `${API}/games`;
+	const method = id ? 'PUT' : 'POST';
+	const res = await fetch(url, {
+		method,
+		headers: { 'Content-Type': 'application/json' },
+		body: JSON.stringify(payload),
+	});
+	if (!res.ok) throw new Error(await res.text());
+}
+
+export async function saveGameFromBgg(payload) {
+	const res = await fetch(`${API}/games/from-bgg`, {
+		method: 'POST',
+		headers: { 'Content-Type': 'application/json' },
+		body: JSON.stringify(payload),
+	});
+	if (!res.ok) throw new Error(await res.text());
+}
+
+export async function deleteGame(id, onDeleted) {
+	const res = await fetch(`${API}/games/${id}`, { method: 'DELETE' });
+	if (!res.ok) throw new Error(await res.text());
+	if (onDeleted) await onDeleted();
+}
+
+export async function searchBgg(q) {
+	const res = await fetch(`${API}/bgg/search?query=${encodeURIComponent(q)}`);
+	return res.json();
+}
+
+export async function fetchBggPreview(bggId) {
+	const res = await fetch(`${API}/bgg/game/${bggId}/preview`);
+	if (!res.ok) return null;
+	return res.json();
+}
