+++ frontend/vite.config.js
@@ -1,0 +1,16 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  root: '.',
+  publicDir: 'public',
+  build: {
+    outDir: 'dist',
+    emptyOutDir: true,
+  },
+  server: {
+    port: 5173,
+    proxy: {
+      '/api': 'http://localhost:5074',
+    },
+  },
+})
